Implemented different filters for sampling pixels

+++ src/color.rs
@@ -6,8 +6,8 @@ use glam::Vec3A;
 pub type Color = Vec3A;
 
 
-pub fn to_rgb(pixel_color: Color, samples_per_pixel: u32) -> image::Rgb<u8> {
-    let scale: f32 = 1.0 / (samples_per_pixel as f32);
+pub fn to_rgb(pixel_color: Color, samples_per_pixel: f32) -> image::Rgb<u8> {
+    let scale: f32 = 1.0 / samples_per_pixel;
     let rgb: Color = (pixel_color * scale).powf(0.5).clamp(Vec3A::new(0.0, 0.0, 0.0), Vec3A::new(0.999, 0.999, 0.999));
     image::Rgb([
         (256.0 * rgb.x) as u8,
@@ -24,7 +24,7 @@ mod tests {
     #[test]
     fn test_to_rgb() -> Result<(), std::fmt::Error> {
         let c: Color = Color::new(0.5, 1.0, 0.0);
-        to_rgb(c, 1);
+        to_rgb(c, 1.0);
         Ok(())
     }
 }
